+++ src/buffer/channel.rs
@@ -508,7 +508,6 @@ mod nick_list {
         theme: &'a Theme,
     ) -> Element<'a, Message> {
         let nicklist_config = &config.buffer.channel.nicklist;
-        let nickname_config = &config.buffer.nickname;
 
         let width = match nicklist_config.width {
             Some(width) => width,
@@ -517,7 +516,7 @@ mod nick_list {
                     .into_iter()
                     .flatten()
                     .map(|user| {
-                        user.display(nickname_config.show_access_levels, None)
+                        user.display(nicklist_config.show_access_levels, None)
                             .chars()
                             .count()
                     })
@@ -530,7 +529,7 @@ mod nick_list {
 
         let content = column(users.into_iter().flatten().map(|user| {
             let content = selectable_text(
-                user.display(nickname_config.show_access_levels, None),
+                user.display(nicklist_config.show_access_levels, None),
             )
             .font_maybe(
                 theme::font_style::nickname(theme, false).map(font::get),